update
Adding scene loading of the in-scene placed NetworkObjects so the client can disconnect and reconnect multiple times in the same runtime instance. Also added NetworkObject clean up on the client side since scene management is disabled (i.e. disconnecting will leave behind artifacts if the active scene is not the same scene that the in-scene placed NetworkObjects were located in).